Move all Module-related classes to the same header file (#1564)
