+++ libs/vgc/ui/CMakeLists.txt
@@ -46,8 +46,6 @@ vgc_add_library(ui
         messagedialog.h
         modifierkey.h
         module.h
-        modulecontext.h
-        modulemanager.h
         mousebutton.h
         mouseevent.h
         numberedit.h
@@ -116,8 +114,6 @@ vgc_add_library(ui
         messagedialog.cpp
         modifierkey.cpp
         module.cpp
-        modulecontext.cpp
-        modulemanager.cpp
         mousebutton.cpp
         mouseevent.cpp
         numberedit.cpp
